Stub out File.exists? for refresh token
So that the oauth refresh token get's generated correctly.
There is an issue where rspec stub stubs all calls to File.exists?
so there is a catch all first that delegates the call to File.exist?
This isn't a nice way of handling it but I can't get a better one to
work.